requirements.txt
# =============================================================================
# Spinal ZF Segmentation Demo - Requirements
# =============================================================================
# For Streamlit Community Cloud deployment
# 
# IMPORTANT NOTES:
# - Use opencv-python-headless (NOT opencv-python) for cloud compatibility
# - Pin versions only when necessary for stability
# - Keep dependencies minimal for faster deployment
# =============================================================================

# Core Framework
streamlit>=1.28.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python-headless>=4.8.0

# Scientific Computing
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Utilities
watchdog>=3.0.0

# =============================================================================
# Optional Dependencies (uncomment if needed)
# =============================================================================
# For advanced explainability methods:
# captum>=0.6.0

# For performance profiling:
# psutil>=5.9.0

# For image format support:
# imageio>=2.31.0
